pyproject.toml
[project]
name = "IoTa"
version = "1.0.0"
description = "IoTa - an IoT data pipeline for Agriculture."
readme = "README.md"
authors = [
    { name = "David Taylor", email = "david.a.taylor@dpi.nsw.gov.au" },
    { name = "Ben Sefton", email = "ben.sefton@dpi.nsw.gov.au" }
]
requires-python = "==3.10.18"
dependencies = [
    "backoff==2.2.1",
    "fastapi==0.116.0",
    "flask==3.1.1",
    "folium==0.20.0",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "httpx==0.28.1",
    "orjson==3.10.18",
    "paho-mqtt==2.1.0",
    "pandas==2.2.1",
    "pika==1.3.2",
    "psycopg2-binary==2.9.3",
    "pytest==7.4.0",
    "python-dotenv==0.20.0",
    "ujson==5.10.0",
    "uvicorn==0.35.0",
    "watchgod==0.8.2",
    "Werkzeug==3.1.3",
]
